Move ServerMonitorAck to appropriate namespace
